LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/controller-on-reload.https-expected.txt



FAIL controller is set upon reload after registration assert_equals: controller should be null until the document is reloaded expected null but got object "[object ServiceWorker]"
